Stop and Searches without location

Returns details of stop and searches that cannot be mapped to a particular location. Note that the police force must be specified. For all stop and searches carried out by a police force, use ukc_stop_search_force().

ukc_stop_search_no_location(force, date = NULL)

Arguments

force

A string containing the name of the police force to return data for. Must be specified, and is not case sensitive. See ukc_forces() for details.

date

The year and month in "YYYY-MM" form. Also accepts dates in formats that can be coerced to Date class with as.Date(). If NULL, the latest available month will be returned. Note that not all forces will necessarily have reported stop and search data for the latest month; check availability with the ukc_available() function.

Value

A tibble with details of stop and searches without a specific location.

See also

Examples

if (FALSE) {
ss_no_location <- ukc_stop_search_no_location(force = "city-of-london")
}