Street level crime

Street level crime

ukc_street_crime(lat, lng, date = NULL, crime_category = NULL)

Arguments

lat

Latitude. Accepts a single value or a vector of values to create a custom polygon.

lng

Longitude. Accepts a single value or a vector of values to create a custom polygon.

date

The year and month in "YYYY-MM" form. If NULL, latest available month will be returned.

crime_category

The category of crime to return. Defaults to returning all crimes. See ukc_crime_category() for details.

Value

A tibble with details of street crimes.

Details

lat and lng must be the same length.

Examples

if (FALSE) {
crime <- ukc_street_crime(
  lat = 51.5, lng = -0.6,
  crime_category = "bicycle-theft"
)

crime_poly <- ukc_street_crime(
  lat = c(52.268, 52.794, 52.130),
  lng = c(0.543, 0.238, 0.478)
)
}