ukp_crime_poly

Extract the crime areas within a polygon

ukp_crime_poly(poly, date = NULL, ...)

Arguments

poly

The lat/lng pairs which define the boundary of the custom area. If a custom area contains more than 10,000 crimes, the API will return a 503 status code. The poly parameter is formatted in lat/lng pairs, separated by colons: [lat],[lng]:[lat],[lng]:[lat],[lng]. The first and last coordinates need not be the same — they will be joined by a straight line once the request is made.

date,

Optional. (YYY-MM), limit results to a specific month. The latest month will be shown by default. e.g. date = "2013-01"

...

further arguments passed to or from other methods. For example, verbose option can be added with ukp_api("call", config = httr::verbose()). See more in ?httr::GET documentation (https://cran.r-project.org/web/packages/httr/) and (https://cran.r-project.org/web/packages/httr/vignettes/quickstart.html).

Note

further documentation here: https://data.police.uk/docs/method/crime-street/

Examples

library(ukpolice)

# with 3 points
ukp_data_poly_3 <- ukp_crime_poly(
poly = c("52.268, 0.543:52.794,0.238:52.130,0.478")
)

#> No encoding supplied: defaulting to UTF-8.

head(ukp_data_poly_3)

#> # A tibble: 6 × 13
#>                category context       id location_subtype location_type
#>                   <chr>   <chr>    <chr>            <chr>         <chr>
#> 1 anti-social-behaviour         51259842                          Force
#> 2 anti-social-behaviour         51257791                          Force
#> 3 anti-social-behaviour         51006452                          Force
#> 4 anti-social-behaviour         51011212                          Force
#> 5 anti-social-behaviour         51009892                          Force
#> 6 anti-social-behaviour         51258293                          Force
#> # ... with 8 more variables: latitude <chr>, longitude <chr>, street_id <chr>,
#> #   street_name <chr>, date <chr>, persistent_id <chr>, outcome_category <chr>,
#> #   outcome_date <chr>

# with 4 points
ukp_data_poly_4 <- ukp_crime_poly(
poly = c("52.268, 0.543:52.794,0.238:52.130,0.478:52.000,0.400")
)

#> No encoding supplied: defaulting to UTF-8.

head(ukp_data_poly_4)

#> # A tibble: 6 × 13
#>                category context       id location_subtype location_type
#>                   <chr>   <chr>    <chr>            <chr>         <chr>
#> 1 anti-social-behaviour         51259842                          Force
#> 2 anti-social-behaviour         51257791                          Force
#> 3 anti-social-behaviour         51006452                          Force
#> 4 anti-social-behaviour         51011212                          Force
#> 5 anti-social-behaviour         51009892                          Force
#> 6 anti-social-behaviour         51258293                          Force
#> # ... with 8 more variables: latitude <chr>, longitude <chr>, street_id <chr>,
#> #   street_name <chr>, date <chr>, persistent_id <chr>, outcome_category <chr>,
#> #   outcome_date <chr>