Find the assoiate neighbourhood code with a long/lat string

ukp_neighbourhood_location takes a longitude/latitude string and finds the associated neighbourhood code associated with that. Read more here https://data.police.uk/docs/method/neighbourhood-locate/

ukp_neighbourhood_location(lat_long, ...)

Arguments

lat_long
a character string of latitude and longitude

...
further arguments passed to or from other methods. For example, verbose option can be added with ukp_api("call", config = httr::verbose()). See more in ?httr::GET documentation (https://cran.r-project.org/web/packages/httr/) and (https://cran.r-project.org/web/packages/httr/vignettes/quickstart.html).

Value

a tibble