Extract crime areas within a polygon

Extract crime areas within a polygon

ukp_crime_poly(poly_df, date = NULL, ...)

Arguments

poly_df

dataframe containing the lat/lng pairs which define the boundary of the custom area. If a custom area contains more than 10,000 crimes, the API will return a 503 status code. ukp_crime_poly converts the dataframe into lat/lng pairs, separated by colons: [lat],[lng]:[lat],[lng]:[lat],[lng]. The first and last coordinates need not be the same — they will be joined by a straight line once the request is made.

date,

Optional. (YYY-MM), limit results to a specific month. The latest month will be shown by default. e.g. date = "2013-01"

...

further arguments passed to or from other methods. For example, verbose option can be added with ukp_api("call", config = httr::verbose()). See more in ?httr::GET documentation https://cran.r-project.org/web/packages/httr/ and https://cran.r-project.org/web/packages/httr/vignettes/quickstart.html.

Note

further documentation here: https://data.police.uk/docs/method/crime-street/

Examples

library(ukpolice)

# with 3 points
poly_df_3 = data.frame(lat = c(52.268, 52.794, 52.130),
                       long = c(0.543, 0.238, 0.478))

ukp_data_poly_3 <- ukp_crime_poly(poly_df_3)

#> No encoding supplied: defaulting to UTF-8.

head(ukp_data_poly_3)

#> # A tibble: 6 × 12
#>                category persistent_id    date      lat     long street_id
#>                   <chr>         <chr>   <chr>    <dbl>    <dbl>     <chr>
#> 1 anti-social-behaviour               2016-11 52.32107 0.442209   1141565
#> 2 anti-social-behaviour               2016-11 52.30487 0.494963   1141325
#> 3 anti-social-behaviour               2016-11 52.34834 0.466663   1142004
#> 4 anti-social-behaviour               2016-11 52.30525 0.490876   1141332
#> 5 anti-social-behaviour               2016-11 52.30889 0.499972   1141317
#> 6 anti-social-behaviour               2016-11 52.30487 0.494963   1141325
#> # ... with 6 more variables: street_name <chr>, context <chr>, id <chr>,
#> #   location_type <chr>, location_subtype <chr>, outcome_status <chr>

# with 4 points
poly_df_4 = data.frame(lat = c(52.268, 52.794, 52.130, 52.000),
                       long = c(0.543,  0.238,  0.478,  0.400))
ukp_data_poly_4 <- ukp_crime_poly(poly_df = poly_df_4)

#> No encoding supplied: defaulting to UTF-8.

head(ukp_data_poly_4)

#> # A tibble: 6 × 12
#>                category persistent_id    date      lat     long street_id
#>                   <chr>         <chr>   <chr>    <dbl>    <dbl>     <chr>
#> 1 anti-social-behaviour               2016-11 52.30889 0.499972   1141317
#> 2 anti-social-behaviour               2016-11 52.08170 0.444165   1137072
#> 3 anti-social-behaviour               2016-11 52.37459 0.480496   1142487
#> 4 anti-social-behaviour               2016-11 52.30487 0.494963   1141325
#> 5 anti-social-behaviour               2016-11 52.34309 0.405690    564425
#> 6 anti-social-behaviour               2016-11 52.35456 0.420852   1142016
#> # ... with 6 more variables: street_name <chr>, context <chr>, id <chr>,
#> #   location_type <chr>, location_subtype <chr>, outcome_status <chr>